sut/application/requirements.txt
fastapi

uvicorn[standard]
gunicorn

prometheus-fastapi-instrumentator

structlog

starlette-exporter

opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-grpc

pyroscope-io

psutil